Add NotificationSettings component
